Fixes PR comments:
- Adds more tests to cover some error cases when creates a composed system
- Adds empty bays to mock of available targets of SPT
- Moves to another file all test of "create a composed system"
