+++ oneview_redfish_toolkit/tests/blueprints/test_computer_system.py
@@ -15,13 +15,10 @@
 # under the License.
 
 # Python libs
-import copy
 import json
 from unittest import mock
 
 # 3rd party libs
-from unittest.mock import call
-
 from flask_api import status
 from hpOneView.exceptions import HPOneViewException
 
@@ -92,29 +89,6 @@ class TestComputerSystem(BaseFlaskTest):
                 'oneview_redfish_toolkit/mockups/redfish/ComputerSystem.json'
         ) as f:
             self.computer_system_mockup = json.load(f)
-
-        with open(
-                'oneview_redfish_toolkit/mockups/redfish/'
-                'PostToComposeSystem.json'
-        ) as f:
-            self.data_to_create_system = json.load(f)
-
-        with open(
-                'oneview_redfish_toolkit/mockups/oneview/'
-                'ServerProfileTemplate.json'
-        ) as f:
-            self.server_profile_template = json.load(f)
-
-        with open(
-                'oneview_redfish_toolkit/mockups/oneview/'
-                'Drives.json'
-        ) as f:
-            self.drives = json.load(f)
-
-        self.sh_id = "30303437-3034-4D32-3230-313131324752"
-        self.spt_id = "1f0ca9ef-7f81-45e3-9d64-341b46cf87e0"
-        self.drive1_id = "e11dd3e0-78cd-47e8-bacb-9813f4bb58a8"
-        self.drive2_id = "53bd734f-19fe-42fe-a8ef-3f1a83b4e5c1"
 
     @mock.patch.object(computer_system, 'g')
     def test_get_computer_system_not_found(self, g):
@@ -607,296 +581,3 @@ class TestComputerSystem(BaseFlaskTest):
         self.assertEqual(
             status.HTTP_500_INTERNAL_SERVER_ERROR, response.status_code)
         self.assertEqual("application/json", response.mimetype)
-
-    @mock.patch.object(computer_system, 'g')
-    def test_create_system(self, g):
-        """Tests create a redfish System with Network, Storage and Server"""
-
-        with open(
-                'oneview_redfish_toolkit/mockups/oneview/'
-                'ServerProfileBuiltFromTemplateToCreateASystem.json'
-        ) as f:
-            expected_server_profile_built = json.load(f)
-
-        g.oneview_client.server_hardware.get.side_effect = [
-            self.server_hardware,
-            self.not_found_error,
-            self.not_found_error,
-            self.not_found_error
-        ]
-        g.oneview_client.server_profile_templates.get.side_effect = [
-            self.not_found_error,
-            self.server_profile_template,
-            self.not_found_error,
-            self.not_found_error,
-            self.server_profile_template
-        ]
-        g.oneview_client.index_resources.get.side_effect = [
-            self.not_found_error,
-            self.not_found_error,
-            self.drives[0],
-            self.drives[1]
-        ]
-        g.oneview_client.server_profiles.create.return_value = \
-            self.server_profile
-
-        response = self.client.post(
-            "/redfish/v1/Systems/",
-            data=json.dumps(self.data_to_create_system),
-            content_type='application/json')
-
-        self.assertEqual(status.HTTP_201_CREATED, response.status_code)
-        self.assertEqual("application/json", response.mimetype)
-        self.assertIn(
-            "/redfish/v1/Systems/" + self.server_profile["uuid"],
-            response.headers["Location"]
-        )
-        g.oneview_client.server_hardware.get.assert_has_calls(
-            [
-                call(self.sh_id),
-                call(self.spt_id),
-                call(self.drive1_id),
-                call(self.drive2_id)
-            ]
-        )
-        g.oneview_client.server_profile_templates.get.assert_has_calls(
-            [
-                call(self.sh_id),
-                call(self.spt_id),
-                call(self.drive1_id),
-                call(self.drive2_id),
-                call(self.spt_id)
-            ]
-        )
-        g.oneview_client.index_resources.get.assert_has_calls(
-            [
-                call('/rest/drives/' + self.sh_id),
-                call('/rest/drives/' + self.spt_id),
-                call('/rest/drives/' + self.drive1_id),
-                call('/rest/drives/' + self.drive2_id)
-            ]
-        )
-
-        g.oneview_client.server_profiles.create.assert_called_with(
-            expected_server_profile_built
-        )
-
-    @mock.patch.object(computer_system, 'g')
-    def test_create_system_when_request_content_is_wrong(self, g):
-        """Tests trying create a redfish System without Links"""
-
-        data_to_send = {
-            "Name": "Composed System Without Links"
-        }
-
-        response = self.client.post("/redfish/v1/Systems/",
-                                    data=json.dumps(data_to_send),
-                                    content_type='application/json')
-
-        self.assertEqual(
-            status.HTTP_400_BAD_REQUEST, response.status_code)
-        self.assertEqual("application/json", response.mimetype)
-        g.oneview_client.server_hardware.get.assert_not_called()
-        g.oneview_client.server_profile_templates.get.assert_not_called()
-        g.oneview_client.index_resources.get.assert_not_called()
-        g.oneview_client.server_profiles.create.assert_not_called()
-
-    @mock.patch.object(computer_system, 'g')
-    def test_create_system_when_request_content_did_not_have_compute(self, g):
-        """Tests trying create a redfish System without Compute Resource"""
-
-        g.oneview_client.server_hardware.get.side_effect = [
-            self.not_found_error,
-            self.not_found_error,
-            self.not_found_error,
-            self.not_found_error
-        ]
-
-        response = self.client.post(
-            "/redfish/v1/Systems/",
-            data=json.dumps(self.data_to_create_system),
-            content_type='application/json')
-
-        self.assertEqual(
-            status.HTTP_400_BAD_REQUEST, response.status_code)
-        self.assertIn(
-            "Should have a Compute Resource Block",
-            response.data.decode()
-        )
-        self.assertEqual("application/json", response.mimetype)
-        g.oneview_client.server_hardware.get.assert_has_calls(
-            [
-                call(self.sh_id),
-                call(self.spt_id),
-                call(self.drive1_id),
-                call(self.drive2_id)
-            ]
-        )
-        g.oneview_client.server_profile_templates.get.assert_not_called()
-        g.oneview_client.index_resources.get.assert_not_called()
-        g.oneview_client.server_profiles.create.assert_not_called()
-
-    @mock.patch.object(computer_system, 'g')
-    def test_create_system_when_request_content_did_not_have_network(self, g):
-        """Tests trying create a redfish System without Network Resource"""
-
-        g.oneview_client.server_hardware.get.side_effect = [
-            self.server_hardware,
-            self.not_found_error,
-            self.not_found_error,
-            self.not_found_error
-        ]
-
-        g.oneview_client.server_profile_templates.get.side_effect = [
-            self.not_found_error,
-            self.not_found_error,
-            self.not_found_error,
-            self.not_found_error
-        ]
-
-        response = self.client.post(
-            "/redfish/v1/Systems/",
-            data=json.dumps(self.data_to_create_system),
-            content_type='application/json')
-
-        self.assertEqual(
-            status.HTTP_400_BAD_REQUEST, response.status_code)
-        self.assertIn(
-            "Should have a valid Network Resource Block",
-            response.data.decode()
-        )
-        self.assertEqual("application/json", response.mimetype)
-        g.oneview_client.server_hardware.get.assert_has_calls(
-            [
-                call(self.sh_id),
-                call(self.spt_id),
-                call(self.drive1_id),
-                call(self.drive2_id)
-            ]
-        )
-        g.oneview_client.server_profile_templates.get.assert_has_calls(
-            [
-                call(self.sh_id),
-                call(self.spt_id),
-                call(self.drive1_id),
-                call(self.drive2_id)
-            ]
-        )
-        g.oneview_client.index_resources.get.assert_not_called()
-        g.oneview_client.server_profiles.create.assert_not_called()
-
-    @mock.patch.object(computer_system, 'g')
-    def test_create_system_when_request_content_did_not_have_a_valid_network(
-            self, g):
-        """Tests trying create a redfish System with a invalid Network"""
-
-        g.oneview_client.server_hardware.get.side_effect = [
-            self.server_hardware,
-            self.not_found_error,
-            self.not_found_error,
-            self.not_found_error
-        ]
-
-        g.oneview_client.server_profile_templates.get.side_effect = [
-            self.not_found_error,
-            self.server_profile_template,
-            self.not_found_error,
-            self.not_found_error
-        ]
-
-        data_to_send = copy.copy(self.data_to_create_system)
-        data_to_send["Id"] = "123456789"
-
-        response = self.client.post(
-            "/redfish/v1/Systems/",
-            data=json.dumps(data_to_send),
-            content_type='application/json')
-
-        self.assertEqual(
-            status.HTTP_400_BAD_REQUEST, response.status_code)
-        self.assertIn(
-            "Should have a valid Network Resource Block",
-            response.data.decode()
-        )
-        self.assertEqual("application/json", response.mimetype)
-        g.oneview_client.server_hardware.get.assert_has_calls(
-            [
-                call(self.sh_id),
-                call(self.spt_id),
-                call(self.drive1_id),
-                call(self.drive2_id)
-            ]
-        )
-        g.oneview_client.server_profile_templates.get.assert_has_calls(
-            [
-                call(self.sh_id),
-                call(self.spt_id),
-                call(self.drive1_id),
-                call(self.drive2_id)
-            ]
-        )
-        g.oneview_client.index_resources.get.assert_not_called()
-        g.oneview_client.server_profiles.create.assert_not_called()
-
-    @mock.patch.object(computer_system, 'g')
-    def test_create_system_when_request_content_did_not_have_storage(self, g):
-        """Tests create a redfish System without Storage. Should works well"""
-
-        g.oneview_client.server_hardware.get.side_effect = [
-            self.server_hardware,
-            self.not_found_error,
-            self.not_found_error,
-            self.not_found_error
-        ]
-        g.oneview_client.server_profile_templates.get.side_effect = [
-            self.not_found_error,
-            self.server_profile_template,
-            self.not_found_error,
-            self.not_found_error,
-            self.server_profile_template
-        ]
-        g.oneview_client.index_resources.get.side_effect = [
-            self.not_found_error,
-            self.not_found_error,
-            self.not_found_error,
-            self.not_found_error
-        ]
-        g.oneview_client.server_profiles.create.return_value = \
-            self.server_profile
-
-        response = self.client.post(
-            "/redfish/v1/Systems/",
-            data=json.dumps(self.data_to_create_system),
-            content_type='application/json')
-
-        self.assertEqual(status.HTTP_201_CREATED, response.status_code)
-        self.assertEqual("application/json", response.mimetype)
-        self.assertIn(
-            "/redfish/v1/Systems/" + self.server_profile["uuid"],
-            response.headers["Location"]
-        )
-        g.oneview_client.server_hardware.get.assert_has_calls(
-            [
-                call(self.sh_id),
-                call(self.spt_id),
-                call(self.drive1_id),
-                call(self.drive2_id)
-            ]
-        )
-        g.oneview_client.server_profile_templates.get.assert_has_calls(
-            [
-                call(self.sh_id),
-                call(self.spt_id),
-                call(self.drive1_id),
-                call(self.drive2_id),
-                call(self.spt_id)
-            ]
-        )
-        g.oneview_client.index_resources.get.assert_has_calls(
-            [
-                call('/rest/drives/' + self.sh_id),
-                call('/rest/drives/' + self.spt_id),
-                call('/rest/drives/' + self.drive1_id),
-                call('/rest/drives/' + self.drive2_id)
-            ]
-        )
